fix: change parameter type from uint32_t to size_t in mem_set function for consistency

+++ kernel/mem.h
@@ -5,4 +5,4 @@
 
 void* memmove(void* dest, const void* src, size_t n);
 void memory_copy(void *src, void *dest, size_t nbytes);
-void mem_set(void *dst, uint8_t val, uint32_t count);
+void mem_set(void *dst, uint8_t val, size_t count);
